devel/py-kivy: Pull py-kivy 1.7.2 from OpenBSD-wip github repository.
This port is untested and was being developed by Edd Barrett,
but then it was deleted as no progress was made on it.

I have been able to compile Kivy 1.11.1 on OpenBSD separately.
Hopefully I will be able to bring Kivy up to at least 1.11.0.
The latest release on GitHub from June 01, 2019 is Version 1.11.0.