devel/py-kivy: Update port to Kivy 1.11.1.
Makefile had a major overhaul, with most work in WANTLIB and
LIB_DEPENDS. Not sure if gstreamer plugins are needed or not.

Hardcoded Python3 FLAVOR for now. Obviously not appropriate if
this was submitted to the OpenBSD Ports maintainers.

Getting distribution tarball from Github. Seems to be authoritative.

There are RUN_DEPENDS on xclip and xsel for using the clipboard.

Patches are hacks, but they work. Hardcoded /usr/local, _clockid is 3
for CLOCK_MONOTONIC in OpenBSD, and forcing OpenBSD to pretend to be Linux.

setup.py had a hard time finding X11 header files, so forced
/usr/X11R6/include and /usr/X11R6/lib to be found so that Cython
can compile down the Python graphics code.

No testing was done, and the test harness in the code has been disabled.

Examples are installed in /usr/local/share/kivy-examples.

Package is about 14M in size. Not too bad.