devel/mips-elf: GCC and Binutils toolchain for MIPS 32-bit.
GCC is version 4.8.1; Binutils is version 2.27.

This toolchain's initial purpose is to build RetroBSD on OpenBSD.
RetroBSD is a port of 2.11 BSD to MIPS-based PIC32 processors.
The current state of the RetroBSD code (as of this commit) can
compile on Ubuntu with the gcc 4.8.1 and binutils 2.27.
The VirtualMIPS simulator can run the compiled code on Ubuntu.
It is this port's intention to be able to compile and run RetroBSD
on OpenBSD, using OpenBSD as the cross-compilation host.