emulators/qemu-arm: Forked QEMU for embedded Arm Cortex-M.
Version 2.8.0-6-20190517 released on April 24, 2019 on GitHub at
github.com/ilg-archived/qemu and is a fork of Version 2.8.0 QEMU,
and includes many changes that are specific to the Arm Cortex-M.

WXNEEDED needs to be enabled, since TCG is incompatible with the
benefits of W^X protection. This fixes a difficult to diagnose
segmentation fault down in the TCG code.

This port is based off the standard emulators/qemu port, but does
not build all the other targets. This is only an Arm Cortex-M port.