devel/arm-none-eabi/gcc-linaro: Enable arm rmprofile multilib.
Small revision bump (to '1', since OpenBSD Ports is at '0') that
enables the 'rmprofile' multilib for the 'arm' port FLAVOR.
This splits out separate libgcc.a libraries for all the differing
Cortex-M and Cortex-R architecture profiles (i.e., v6-m, v7-m, v7e-m)
with both soft and hard floating point support.

'make update-plist' suggested to make the library archives '@static-lib',
but other than that, no change to the 'aarch64' port FLAVOR.

The PLIST and PFRAG.* packing lists were hand-massaged, as update-plist
has a hard time dealing with the FLAVORS for this port.

Note that the current -newlib cannot be built by this new multilib gcc,
as 2.2.0-1 is old and does not have adequate ARMv6-M or ARMv8-M support.